requirements.txt
annotated-types==0.6.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
paho-mqtt==1.6.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
